+++ src/Sulu/Bundle/MediaBundle/Resources/public/js/components/media-selection/main.js
@@ -339,7 +339,7 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
         bindCustomEvents = function() {
             this.sandbox.on('husky.tabs.overlaymedia-selection.' + this.options.instanceName + '.add.initialized', function() {
                 startOverlayLoader.call(this);
-                this.collections.fetch({
+                this.collections.fetchSorted('title', {
                     success: function(collections) {
                         this.collectionArray = collections.toJSON();
                         stopOverlayLoader.call(this);
@@ -420,7 +420,8 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
         startGridGroup = function() {
             // TODO remove high limit and implement automatic reload pagination
             var gridUrl = '/admin/api/media?limit=99999&' +
-                (this.options.types != '' ? 'types=' + this.options.types + '&' : '');
+                (this.options.types != '' ? 'types=' + this.options.types + '&' : '') +
+                'collection='; // TODO remove this quickfix only show specific media from collections
 
             this.sandbox.start([
                 {