+++ src/Sulu/Bundle/SearchBundle/Command/ReindexCommand.php
@@ -11,15 +11,16 @@
 namespace Sulu\Bundle\SearchBundle\Command;
 
 use Jackalope\Query\Row;
-use Jackalope\Session;
-use Sulu\Bundle\SearchBundle\LocalizedSearchManager\LocalizedSearchManagerInterface;
+use Massive\Bundle\SearchBundle\Search\SearchManagerInterface;
+use PHPCR\SessionInterface;
 use Sulu\Component\Content\Mapper\ContentMapperInterface;
 use Sulu\Component\Content\Structure;
+use Sulu\Component\Util\SuluNodeHelper;
 use Sulu\Component\Webspace\Manager\WebspaceManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
+use Symfony\Component\Config\Definition\Exception\Exception;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Sulu\Component\Content\Mapper\Translation\MultipleTranslatedProperties;
 
 class ReindexCommand extends ContainerAwareCommand
 {
@@ -37,13 +38,14 @@ EOT
     public function execute(InputInterface $input, OutputInterface $output)
     {
         $container = $this->getContainer();
-        /** @var Session $session */
+
+        /** @var SessionInterface $session */
         $session = $container->get('sulu.phpcr.session')->getSession();
 
         /** @var ContentMapperInterface $contentMapper */
         $contentMapper = $container->get('sulu.content.mapper');
 
-        /** @var LocalizedSearchManagerInterface $searchManager */
+        /** @var SearchManagerInterface $searchManager */
         $searchManager = $container->get('massive_search.search_manager');
 
         /** @var WebspaceManagerInterface $webspaceManager */
@@ -53,7 +55,7 @@ EOT
         $webspacePrefix = $container->getParameter('sulu.content.node_names.base');
         $tempName = $container->getParameter('sulu.content.node_names.temp');
 
-        $sql2 = 'SELECT * FROM [nt:unstructured] AS a WHERE [jcr:mixinTypes] = "sulu:content"';
+        $sql2 = 'SELECT * FROM [nt:unstructured] AS a WHERE [jcr:mixinTypes] = "sulu:page"';
 
         $queryManager = $session->getWorkspace()->getQueryManager();
 
@@ -61,18 +63,17 @@ EOT
 
         $res = $query->execute();
 
-        $multiTranslatedProperties = new MultipleTranslatedProperties(array(), 'i18n', '');
+        /** @var SuluNodeHelper $nodeHelper */
+        $nodeHelper = $container->get('sulu.util.node_helper');
 
         /** @var Row $row */
         foreach ($res->getRows() as $row) {
             $node = $row->getNode('a');
 
-            // Evil 1: This should be encapsulated in a domain object Structure
-            //         which we havn't implemented yet.
-            $locales = $multiTranslatedProperties->getLanguagesForNode($node);
+            $locales = $nodeHelper->getLanguagesForNode($node);
             foreach ($locales as $locale) {
 
-                // Evil 2: Also should be encapsulated.
+                // Evil: Should be encapsulated.
                 if (!preg_match('{/' . $webspacePrefix . '/(.*?)/(.*?)(/.*)*$}', $node->getPath(), $matches)) {
                     $output->writeln(
                         sprintf('<error> - Could not determine webspace for </error>: %s', $node->getPath())
@@ -84,16 +85,25 @@ EOT
                 if ($tempName !== $matches[2] && $webspaceManager->findWebspaceByKey($webspaceKey) !== null) {
                     $structure = $contentMapper->load($node->getIdentifier(), $webspaceKey, $locale);
 
-                    if ($structure->getNodeState() === Structure::STATE_PUBLISHED) {
+                    try {
+                        if ($structure->getNodeState() === Structure::STATE_PUBLISHED) {
+                            $output->writeln(
+                                '  [+] <comment>Indexing published page (locale: ' . $locale . ')</comment>: ' .
+                                $node->getPath()
+                            );
+                            $searchManager->index($structure, $locale);
+                        } else {
+                            $output->writeln(
+                                '  [-] <comment>De-indexing unpublished page (locale: ' . $locale . ')</comment>: ' .
+                                $node->getPath()
+                            );
+                            $searchManager->deindex($structure, $locale);
+                        }
+                    } catch (\Exception $exc) {
                         $output->writeln(
-                            '  [+] <comment>Indexing published structure (locale: ' . $locale . ')</comment>: ' . $node->getPath()
+                            '  [!] <error>Error indexing or de-indexing page (path: ' . $node->getPath() .
+                            ', locale: ' . $locale . '): ' . $exc->getMessage() . '</error>'
                         );
-                        $searchManager->index($structure, $locale);
-                    } else {
-                        $output->writeln(
-                            '  [-] <comment>De-indexing unpublished structure (locale: ' . $locale . ')</comment>: ' . $node->getPath()
-                        );
-                        $searchManager->deindex($structure, $locale);
                     }
                 }
             }