+++ src/Sulu/Bundle/TestBundle/Testing/PhpcrTestCase.php
@@ -186,12 +186,16 @@ abstract class PhpcrTestCase extends \PHPUnit_Framework_TestCase
             $this->prepareLocalizationFinder();
 
             $this->templateResolver = new TemplateResolver();
-            $nodeHelper = new SuluNodeHelper('i18n', array(
-                'base' => 'cmf',
-                'content' => 'contents',
-                'route' => 'routes',
-                'snippet' => 'snippets',
-            ));
+            $nodeHelper = new SuluNodeHelper(
+                $this->sessionManager->getSession(),
+                'i18n', 
+                array(
+                    'base' => 'cmf',
+                    'content' => 'contents',
+                    'route' => 'routes',
+                    'snippet' => 'snippets',
+                )
+            );
             $cleaner = new PathCleanup();
             $strategy = new TreeStrategy(new PhpcrMapper($this->sessionManager, '/cmf/routes'), $cleaner);