+++ src/Sulu/Bundle/TestBundle/Testing/SuluTestCase.php
@@ -7,6 +7,7 @@ use Doctrine\Common\DataFixtures\Purger\ORMPurger;
 use Doctrine\Common\DataFixtures\Executor\ORMExecutor;
 use Doctrine\Common\DataFixtures\ProxyReferenceRepository;
 use Sulu\Component\Content\Structure;
+use Sulu\Bundle\TestBundle\Kernel\SuluTestKernel;
 
 /**
  * Base test case for functional tests in Sulu
@@ -15,6 +16,47 @@ use Sulu\Component\Content\Structure;
  */
 abstract class SuluTestCase extends BaseTestCase
 {
+    static protected $kernels = array();
+    static protected $currentKernel = 'admin';
+
+    /**
+     * Create a new SuluTestKernel and pass the sulu.context to it.
+     *
+     * {@inheritDoc}
+     *
+     * @throws InvalidArgumentEXception If the found kernel does
+     *   not extend SuluTestKernel
+     */
+    protected static function createKernel(array $options = array())
+    {
+        if (null === static::$class) {
+            static::$class = static::getKernelClass();
+        }
+
+        $kernel = new static::$class(
+            isset($options['environment']) ? $options['environment'] : 'test',
+            isset($options['debug']) ? $options['debug'] : true,
+            isset($options['sulu_context']) ? $options['sulu_context'] : 'admin'
+        );
+
+        if (!$kernel instanceof SuluTestKernel) {
+            throw new \InvalidArgumentException(sprintf(
+                'All Sulu testing Kernel classes must extend SuluTestKernel, "%s" does not',
+                get_class($kernel)
+            ));
+        }
+
+        return $kernel;
+    }
+
+    /**
+     * Close the database connection after the tests finish
+     */
+    public function tearDown()
+    {
+        $this->db('ORM')->getOm()->getConnection()->close();
+    }
+
     /**
      * Return the ID of the test user (which is provided / created
      * by the test_user_provider in this Bundle at runtime)
@@ -30,7 +72,39 @@ abstract class SuluTestCase extends BaseTestCase
     }
 
     /**
+     * Create an authenticated client
+     *
+     * @return Symfony\Bundle\FrameworkBundle\Client
+     */
+    protected function createAuthenticatedClient()
+    {
+        return $this->createClient(
+            array(
+                'environment' => 'dev',
+            ),
+            array(
+                'PHP_AUTH_USER' => 'test',
+                'PHP_AUTH_PW' => 'test',
+            )
+        );
+    }
+
+    /**
+     * Create client for tests on the "website" context
+     *
+     * @return Symfony\Bundle\FrameworkBundle\Client
+     */
+    protected function createWebsiteClient()
+    {
+        return $this->createClient(array(
+            'sulu_context' => 'website',
+            'environment' => 'dev',
+        ));
+    }
+
+    /**
      * Initialize / reset the Sulu PHPCR environment
+     * NOTE: This should use initializers when we implement that feature
      */
     protected function initPhpcr()
     {
@@ -71,7 +145,6 @@ abstract class SuluTestCase extends BaseTestCase
 
     /**
      * Purge the Doctrine ORM database
-        $this->import->resetPackages();
      */
     protected function purgeDatabase()
     {
@@ -93,24 +166,4 @@ abstract class SuluTestCase extends BaseTestCase
         }
     }
 
-    /**
-     * Create an authenticated client
-     *
-     * @return Symfony\Bundle\FrameworkBundle\Client
-     */
-    protected function createAuthenticatedClient()
-    {
-        return $this->createClient(
-            array(),
-            array(
-                'PHP_AUTH_USER' => 'test',
-                'PHP_AUTH_PW' => 'test',
-            )
-        );
-    }
-
-    public function tearDown()
-    {
-        $this->db('ORM')->getOm()->getConnection()->close();
-    }
 }