+++ src/Sulu/Bundle/WebsiteBundle/Resolver/RequestAnalyzerResolver.php
@@ -40,10 +40,14 @@ class RequestAnalyzerResolver implements RequestAnalyzerResolverInterface
      */
     public function resolve(RequestAnalyzerInterface $requestAnalyzer)
     {
+        // determine default locale (if one exists)
+        $defaultLocalization = $requestAnalyzer->getCurrentPortal()->getDefaultLocalization();
+        $defaultLocale = $defaultLocalization ? $defaultLocalization->getLocalization() : null;
+
         return array(
             'request' => array(
                 'webspaceKey' => $requestAnalyzer->getCurrentWebspace()->getKey(),
-                'defaultLocale' => $requestAnalyzer->getCurrentPortal()->getDefaultLocalization()->getLocalization(),
+                'defaultLocale' => $defaultLocale,
                 'locale' => $requestAnalyzer->getCurrentLocalization()->getLocalization(),
                 'portalUrl' => $requestAnalyzer->getCurrentPortalUrl(),
                 'resourceLocatorPrefix' => $requestAnalyzer->getCurrentResourceLocatorPrefix(),
@@ -67,6 +71,7 @@ class RequestAnalyzerResolver implements RequestAnalyzerResolverInterface
             'request' => array(
                 'webspaceKey' => $webspaceKey,
                 'locale' => $locale,
+                'defaultLocale' => $locale,
                 'portalUrl' => $portalUrl,
                 'resourceLocatorPrefix' => '',
                 'resourceLocator' => '',