+++ src/Sulu/Component/Content/Mapper/ContentMapper.php
@@ -53,6 +53,7 @@ use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\Stopwatch\Stopwatch;
 use Sulu\Component\Content\Event\ContentOrderBeforeEvent;
 use Sulu\Component\Util\SuluNodeHelper;
+use PHPCR\PropertyType;
 
 /**
  * Maps content nodes to phpcr nodes with content types and provides utility function to handle content nodes
@@ -207,7 +208,6 @@ class ContentMapper implements ContentMapperInterface
 
         // optional
         $this->stopwatch = $stopwatch;
-
     }
 
     /**
@@ -393,7 +393,20 @@ class ContentMapper implements ContentMapperInterface
                     $path = $this->getUniquePath($path, $node->getParent());
 
                     if ($path) {
+                        // workaround for Jackalope bug referenced in: https://github.com/sulu-cmf/sulu/issues/518
+                        $nextNode = $this->nodeHelper->getNextNode($node);
+
+                        if ($nextNode) {
+                            $nextNodeName = $nextNode->getName();
+                            $nodeName = $node->getName();
+                        }
+
                         $node->rename($path);
+
+                        if ($nextNode) {
+                            $parentNode = $node->getParent();
+                            $parentNode->orderBefore($node->getName(), $nextNodeName);
+                        }
                     }
                     // FIXME refresh session here
                 }
@@ -1221,6 +1234,11 @@ class ContentMapper implements ContentMapperInterface
             $availableLocalization = $localization;
         }
 
+        // if there was no webspace then determine the webspace from the content node path
+        if (null === $webspaceKey) {
+            $webspaceKey = $this->nodeHelper->extractWebspaceFromPath($contentNode->getPath());
+        }
+
         if ($this->stopwatch) {
             $this->stopwatch->stop('contentManager.loadByNode.available-localization');
             $this->stopwatch->start('contentManager.loadByNode.mapping');
@@ -1524,6 +1542,11 @@ class ContentMapper implements ContentMapperInterface
     private function getLocalizedUrlsForPage(Page $page, NodeInterface $node, $webspaceKey, $segmentKey)
     {
         $localizedUrls = array();
+
+        if (null === $webspaceKey) {
+            $webspaceKey = $this->nodeHelper->extractWebspaceFromPath($node->getPath());
+        }
+
         $webspace = $this->webspaceManager->findWebspaceByKey($webspaceKey);
         $property = $page->getPropertyByTagName('sulu.rlp');
         $property = clone $property;
@@ -1539,7 +1562,9 @@ class ContentMapper implements ContentMapperInterface
             $property->setValue(null);
             $contentType->read($node, $translatedProperty, $webspaceKey, $locale, $segmentKey);
 
-            $localizedUrls[$locale] = $property->getValue();
+            if (null !== $property->getValue()) {
+                $localizedUrls[$locale] = $property->getValue();
+            }
         }
 
         return $localizedUrls;
@@ -1548,12 +1573,12 @@ class ContentMapper implements ContentMapperInterface
     /**
      * {@inheritdoc}
      */
-    public function delete($uuid, $webspaceKey)
+    public function delete($uuid, $webspaceKey, $dereference = false)
     {
         $session = $this->getSession();
         $node = $session->getNodeByIdentifier($uuid);
 
-        $this->deleteRecursively($node);
+        $this->deleteRecursively($node, $dereference);
         $session->save();
     }
 
@@ -1627,6 +1652,9 @@ class ContentMapper implements ContentMapperInterface
     }
 
     /**
+     * TODO: Refactor this. This should not effect the global state of the object, this
+     *       should be scoped for each save request.
+     *
      * TRUE dont rename pages on save
      * @param boolean $noRenamingFlag
      * @return $this
@@ -1792,14 +1820,32 @@ class ContentMapper implements ContentMapperInterface
     }
 
     /**
-     * remove node with references (path, history path ...)
+     * Remove node with references (path, history path ...)
+     *
      * @param NodeInterface $node
+     * @param boolean $dereference Remove REFERENCE properties (or property
+     *   values in the case of multi-value) from referencing nodes
      */
-    private function deleteRecursively(NodeInterface $node)
+    private function deleteRecursively(NodeInterface $node, $dereference = false)
     {
         foreach ($node->getReferences() as $ref) {
             if ($ref instanceof \PHPCR\PropertyInterface) {
                 $child = $ref->getParent();
+
+                if ($dereference) {
+                    if ($ref->isMultiple()) {
+                        $values = $ref->getValue();
+                        foreach ($values as $i => $referringNode) {
+                            if ($node->getIdentifier() === $referringNode->getIdentifier()) {
+                                unset($values[$i]);
+                            }
+                        }
+
+                        $ref->getParent()->setProperty($ref->getName(), $values, PropertyType::REFERENCE);
+                    } else {
+                        $ref->remove();
+                    }
+                }
             } else {
                 $child = $ref;
             }
@@ -2020,7 +2066,7 @@ class ContentMapper implements ContentMapperInterface
             $structure = $this->structureManager->getStructure($templateKey);
 
             if (!isset($url)) {
-                $url = $this->getUrl($path, $row, $structure, $webspaceKey, $locale);
+                $url = $this->getUrl($path, $row, $structure, $webspaceKey, $originLocale);
             }
 
             // get url returns false if route is not this language