+++ src/Sulu/Component/Util/SuluNodeHelper.php
@@ -11,8 +11,10 @@
 namespace Sulu\Component\Util;
 
 use PHPCR\NodeInterface;
+use PHPCR\PropertyInterface;
 use PHPCR\Util\PathHelper;
 use Sulu\Component\Content\Structure;
+use PHPCR\SessionInterface;
 
 /**
  * Utility class for extracting Sulu-centric properties from nodes.
@@ -26,19 +28,34 @@ class SuluNodeHelper
     private $languageNamespace;
 
     /**
+     * @var string
+     */
+    private $paths;
+
+    /**
+     * @var SessionInterface
+     */
+    private $session;
+
+    /**
+     * @param SessionInterface $session
      * @param string $languageNamespace
      * @param array $paths Path segments from configuration
      */
-    public function __construct($languageNamespace, $paths)
+    public function __construct(SessionInterface $session, $languageNamespace, $paths)
     {
         $this->languageNamespace = $languageNamespace;
-        $this->paths = array_merge(array(
-            'base' => null,
-            'content' => null,
-            'route' => null,
-            'temp' => null,
-            'snippet' => null
-        ), $paths);
+        $this->paths = array_merge(
+            array(
+                'base' => null,
+                'content' => null,
+                'route' => null,
+                'temp' => null,
+                'snippet' => null
+            ),
+            $paths
+        );
+        $this->session = $session;
     }
 
     /**
@@ -52,7 +69,8 @@ class SuluNodeHelper
     {
         $languages = array();
         foreach ($node->getProperties() as $property) {
-            preg_match('/^' . $this->languageNamespace . ':(.*?)-title/', $property->getName(), $matches);
+            /** @var PropertyInterface $property */
+            preg_match('/^' . $this->languageNamespace . ':([a-zA-Z_]*?)-title/', $property->getName(), $matches);
 
             if ($matches) {
                 $languages[$matches[1]] = $matches[1];
@@ -88,19 +106,21 @@ class SuluNodeHelper
      * Return true if the given node has the given
      * nodeType property (or properties).
      *
-     * The sulu node type is the localname of node types
+     * The sulu node type is the local name of node types
      * with the sulu namespace.
      *
      * Example:
      *   sulu:snippet is the PHPCR node type
      *   snippet is the Sulu node type
      *
-     * @param NodeInterface $nodeTypes
-     * @param string|array $nodeTypes One or more node sulu types
+     * @param NodeInterface $node
+     * @param string|array $suluNodeTypes One or more node sulu types
+     *
+     * @return bool
      */
     public function hasSuluNodeType($node, $suluNodeTypes)
     {
-        foreach ((array) $suluNodeTypes as $suluNodeType) {
+        foreach ((array)$suluNodeTypes as $suluNodeType) {
             if (in_array($suluNodeType, $node->getPropertyValueWithDefault('jcr:mixinTypes', array()))) {
                 return true;
             }
@@ -134,6 +154,7 @@ class SuluNodeHelper
      * @param string $path
      *
      * @return string
+     * @throws \InvalidArgumentException
      */
     public function extractSnippetTypeFromPath($path)
     {
@@ -163,10 +184,65 @@ class SuluNodeHelper
     }
 
     /**
+     * Return the next node of the given node
+     *
+     * @see getSiblingNode
+     */
+    public function getNextNode(NodeInterface $node)
+    {
+        return $this->getSiblingNode($node);
+    }
+
+    /**
+     * Return the previous node of the given node
+     *
+     * @see getSiblingNode
+     */
+    public function getPreviousNode(NodeInterface $node)
+    {
+        return $this->getSiblingNode($node, true);
+    }
+
+    /**
+     * Return either the next or previous sibling of the given node
+     * according to the $previous flag.
+     *
+     * @param NodeInterface $node
+     * @param bool $previous
+     *
+     * @return NodeInterface|null
+     * @throws \RuntimeException
+     */
+    private function getSiblingNode(NodeInterface $node, $previous = false)
+    {
+        $parentNode = $node->getParent();
+        $children = $parentNode->getNodes();
+        $previousNode = null;
+
+        while ($child = current($children)) {
+            if ($child->getPath() === $node->getPath()) {
+                return $previous ? $previousNode : next($children);
+            }
+
+            $previousNode = $child;
+            next($children);
+        }
+
+        throw new \RuntimeException(
+            sprintf(
+                'Could not find node with path "%s" as a child of "%s". This should not happen',
+                $node->getPath(),
+                $parentNode->getPath()
+            )
+        );
+    }
+
+    /**
      * Return the configured named path segment
      *
      * @param string $name Name of path segment
      * @return string The path segment
+     * @throws \InvalidArgumentException
      */
     private function getPath($name)
     {