+++ CHANGELOG.md
@@ -4,16 +4,35 @@ CHANGELOG for Sulu
 * dev-develop
 
     * ENHANCEMENT #539 [AdminBundle] Added validation for iban and vat numbers from the eu
-    * BUGFIX #531 [ContentBundle] Fixed single internal link freeze
-    * ENHANCEMENT #523 [ContentBundle] Prefix contentbundle template path
-    * ENHANCEMENT #526 [WebsiteBundle] Added Template var to resolver (Twig-Template)
+    * HOTFIX      #559 [CoreBundle]    Workaround upstream reg. in DoctrinePHPCRBundle, which causes 
+                                       eager validation of workspace existence.
+    * ENHANCEMENT #523 [All]           Refactored and improved functional tests
+ 
+* 0.11.1 (2014-11-13)
+
+    * HOTFIX      #543 [SearchBundle]  Fixed re-index command
+    * HOTFIX      #551 [SearchBundle]  Switched to test adapter for tests
+    * HOTFIX      #549 [ContentBundle] Fixed page URL fetching for internal links used in snippets
+    * HOTFIX      #512 [MediaBundle]   Only show Media from specific selected Collection
+    * HOTFIX      #550 [MediaBundle]   Deleted Media do not throw Exception when page is saved
+
+* 0.11.0 (2014-11-12)
+
+    * BUGFIX      #540                 Real url in requested language (navigation, ...) for shadows
+    * ENHANCEMENT #523 [ContentBundle] Prefix ContentBundle template path
+    * BUGFIX      #531 [ContentBundle] Fixed single internal link freeze
+    * BUGFIX      #529 [MediaBundle]   Display sorted Collections in overlay
+    * FEATURE     #536 [MediaBundle]   Added Configurable display options for media-selection
     * ENHANCEMENT #361 [WebsiteBundle] Read urls for pages in all languages
+    * ENHANCEMENT #526 [WebsiteBundle] Added Template var to resolver (Twig-Template)
     * ENHANCEMENT #528 [WebsiteBundle] Added memoize service to cache data and use it in twig extension
-    * FEATURE #536 [MediaBundle] Added Configurable display options for media-selection
+    * ENHANCEMENT #524 [ContentBundle] Prefix ContentBundle template path
+    * BUGFIX      #518 [ContentBundle] Ordering of page changed when node is renamed
+    * FEATURE     #511 [SnippetBundle] Ask confirmation when deleting Snippets which are referenced by content
 
 * 0.10.2 (2014-11-07)
 
-  	* HOTFIX #509 [ContentBundle] Fixed cached data bug in smart-content
+    * HOTFIX #509 [ContentBundle] Fixed cached data bug in smart-content
 
 * 0.10.1 (2014-11-04)